Move Error and PlatformError to webauthn:error
