+++ libwebauthn/src/transport/cable/known_devices.rs
@@ -4,9 +4,9 @@ use std::sync::Arc;
 
 use crate::transport::cable::advertisement::await_advertisement;
 use crate::transport::cable::crypto::{derive, KeyPurpose};
-use crate::transport::error::Error;
+use crate::transport::error::TransportError;
 use crate::transport::Device;
-use crate::webauthn::TransportError;
+use crate::webauthn::error::Error;
 use crate::UxUpdate;
 
 use async_trait::async_trait;
